:heavy_minus_sign: Removed AlertDialogMaterial3 file and refactored the usage
The file `AlertDialogMaterial3.kt` was deleted as it was no longer necessary with the adjustments made to the `DefaultAlertDialog.kt` & `InputScreen.kt` files. The deprecated `AlertDialogMaterial3` function was replaced with an `AlertDialog` function in `DefaultAlertDialog.kt`, eliminating the need for the `showDialog` argument. The experimental material API is also deactivated in `InputScreen.kt`. These changes are part of a larger effort to streamline our dialog functions and keep our code cleaner and more efficient.