:test_tube: Added TestFilePicker.kt for UI testing.
A new file, TestFilePicker.kt has been added under the testing directory of the songfinder UI module. This test file allows us to interactively test the file and directory picker functionality of the application. The implementation for the file and directory pickers has been adapted from the sample code provided in the 'darkrockstudios' library repository. A UI to choose file/directory and display the chosen path has been set up.