The docker setup for the project has been restructured. This was necessary to provide a clear separation between the base application and test environments.
The "docker-compose.gradle-test.yml" file has been removed as it was functionally similar to the new "compose.test.yml" file which is specifically designed to be used in a CI environment. It runs tests in a very fresh environment, avoiding possible test contamination.

A new "compose.base.yml" file was introduced, defining two basic services: app and mariadb. The base compose file provides the necessary environment for the application to connect to a mariadb database.

Additionally, a new ".env" file was added for storing and using environment variables, such as database credentials. This avoids having clear-text sensitive data in the docker files. It also allows for easy updating of these details as necessary.