:bug: Switch Kotlin plugin from AllOpen to Spring
The current build was using the Kotlin AllOpen plugin,
which was running into issues of not opening the `@Configuration` classes.
To resolve this,
the build.gradle.kts and libs.versions.toml files were edited
to instead use the Kotlin Spring plugin.
This change should fix the stated error
and enhance integration with Spring framework.