:recycle: Refactor build.gradle.kts files for readability and maintainability
This commit streamlines the structure of the build.gradle.kts files, optimizing for readability and maintainability. Dependencies are now declared in a more centralized manner under songfinder-app's build.gradle.kts file, making them easier to locate and update. Redundancies are removed in the gradle/plugins/convention/compose-desktop-app and gradle/platform/dev-version-constraints modules. The changes also include updated versioning in libs.versions.toml. This shift intends to provide a clear and concise view of the project dependencies and version control.