:heavy_minus_sign: Removal of the Kamel Image dependency from build.gradle.kts
The Kamel Image dependency (libs.dep.composeKamelImage) has been removed from the songfinder-app's build.gradle.kts file. Although this library is a powerful and efficient image-loader library for the compose UI, it was ruled out because it comes with numerous other dependencies that aren't required for the current project, such as Ktor. This move is part of an ongoing initiative to keep the codebase lean and only comprise required dependencies to maintain a faster build time. Hence, it has been commented out in both the build.gradle.kts file and the libs.versions.toml.