:art: Add SearchPanel and FinishMessagePanel to main UI
In this commit, two new Composable functions "SearchPanel" and "FinishMessagePanel" were added to the main UI. The "SearchPanel" is intended to be used when the app is in the searching stage, while "FinishMessagePanel" will be shown when all search tasks are done. These changes are for the enhancement of user interaction: enabling users to conduct searching tasks and receive a feedback message when all tasks are finished.