:sparkles: Add state handling for next line read in IOController
In order to improve the maintainability and scalability of the 'IOController', a private lazy 'currentLineState' has been added to manage state. The existing 'readNextLine()' method has been updated to 'readNextLineToState()' which now utilizes the new state variable for line reads. The change contributes to easier debugging and better handling of I/O operations.