:sparkles: Implement SearchBar in SearchPanel
Extended the functionality of the SearchPanel component by implementing SearchBar for song title input. The SearchBar also includes an icon and tooltip for user convenience and clarity. The song title is fetched from the ioController's currentLineState. Also, added the RealSearchBar invocation to the MainScreen. Further functionalities of backend search are yet to be implemented. This change aims to improve the user interface by offering the ability to easily search for songs by title.