:recycle: Refactor app entry point for spring boot application
In the diff, the entry point for the Spring Boot application is refactored from the `Entrypoint.kt` file to a new file `SongFinderSpringBootApp.kt`. This includes the moving of the `@SpringBootApplication` annotation and the Spring context declaration. This is done to separate concerns, improve modularity and make the context easier to access from anywhere in the application. The `Entrypoint.kt` file now contains only the application launching logic and UI rendering, making it cleaner and easier to navigate.