:recycle: Refactor Bean access method and add progress tooltip
Switched to a more streamlined way of accessing Spring Beans by using the getSpringBean method. This allowed for cleaner and more readable code. In addition, tooltips have been added to the progress bar in the main screen. These tooltips provide an explanation to users about what the progress bar indicates - that is, the number of songs that have been processed versus the total number of songs that will be processed. This aims to enhance user understanding of the process taking place.