:sparkles: Add VocaDbCoroutineRepository interface
This commit adds a new interface, VocaDbCoroutineRepository, in the 'db.repository' package. This interface extends the CrudRepository, ListCrudRepository, and PagingAndSortingRepository interfaces. It will be used as the base repository for our backend services that deal with song information. This modification will allow us to implement database interactions in a more consistent and scalable manner.