:sparkles: Add SongName repository and related tests
Added a new repository 'SongNameRepository'
for handling operations related to SongName in the database.
The repository includes a method
to retrieve all songs by their ID. To ensure functionality,
a respective unit test was also created 'SongNameRepositoryTest'.
This addition is part of the database interaction enhancement for the SongFinder app.