:fire: Add test for SongRepository and refactor repository structure
A method `findByAllPossibleNames` is added to SongRepository with its corresponding test in SongRepositoryTest. This new method allows looking up songs by matching a string to any possible names, improving the usability of search function in the app.

Furthermore, `VocaDbCoroutineRepository` is renamed to `VocaDbRepository` and all of its extensions are updated accordingly. This change is due to the upcoming refactoring that we won't be using Coroutines in the repository layer.