:hammer: Update Pv entity and add PvRepositoryTest
Updated the column name in the Pv entity from 'pv_service' to 'service' for consistency in database naming conventions. Also, added a new test class, PvRepositoryTest, to validate the functionality of retrieving all PVs given a list of song IDs. This change will help in maintaining the code readability and ensuring the correct functionality of the 'findAllBySongIdIn' method.