:memo: Add Builder to SongSearchResult and update fields
The SongSearchResult class has been extended with a Builder class to simplify the construction of instances. This provides a more flexible way of initializing this class. Furthermore, a SongType field was added to the SongSearchResult class to specify the type of song. Also, the publishDate field is now nullable, and the PVInfo now includes a reference to PvsService.