:sparkles: Add ArtistName entity and repository
This commit introduces the ArtistName entity and its corresponding repository. The ArtistName entity represents the names of the artists stored in the database. Meanwhile, the ArtistNameRepository extends the VocaDbRepository and provides a method that retrieves artist names based on artistIds and name languages. This structure is important to enable different functionalities of the SongFinder backend, such as filtering or searching by artist name.