:art: Add Artist related files and test
The commit introduces Artist data structures (Artist, ArtistType, ArtistInSong) that allow to represent and work with Artists data in the application. Also, an ArtistRepository is introduced to explore and manipulate Artists data in the database. The commit also contains a test for the ArtistRepository to ensure it works as expected. These are necessary for handling artist related data within the song-finder app.