:alembic: Updated the search functionality in songfinder-app
In this commit, numerous changes were made to improve the search functionality of the song finder application. Besides, some enhancements were made to the song ID representation.

For Cohesion, a new function 'isOnlineService' was added into the 'Pv.kt' file, which checks if the file is not stored locally.

Furthermore, a type system change was applied to the 'Song.kt' file where the 'id' type was changed from 'Int' to 'Long' for better usage and compatibility.

In the SearchBarController.kt the search functionality has been improved to handle exceptions and better output of search results. The model SongSearchResult.kt has been updated to include 'pvType' in PVInfo, which will help identify the type of 'pv'.

The searchBar UI has additional delay for better user input handling.

Finally, the SongSearchService which was initially mock-up now does real search for songs and added into the search result, it searches by song's title, finds all the PVs related to the songs, and also gets all the artists related to the songs.