:sparkles: Add SearchResultStatusModel and integrate with SearchBarController
A new SearchResultStatusModel class was added to track the search status. It provides two mutable state properties: statusState and resultState, along with methods to update these states. It was integrated into SearchBarController to manage the state of song search operation. This is done right before starting the search and after getting the results or in case an empty input was given. This allows to reactively update the UI and provide responsive user experience.