:sparkles: Added search status indicator to the SearchBar component
This commit adds a CircularProgressIndicator and an Icon for showing when the application is searching or done searching, respectively. This is achieved by adding a new parameter in the SearchBarModel for keeping track of the search status state, and creating a new SearchProgressIndicator function in the SearchBar component that reacts to the changes in the search status state. Additionally, the commit includes minor reformatting and naming adjustments for clarity and consistency.