:zap: Update SearchBar UI and enhance search input responsiveness
Changed the execution of the search function in `SearchBar.kt` from being launched in a CoroutineScope to being launched in a LaunchedEffect, which allows the async task to react to changes in the `value` state. This change improves UI responsiveness to the user's input in the search bar.

Updated searchTextField function to SearchTextField and added TooltipAreaWithCard to the search and done indicators. This gives the user more feedback during the song searching process, especially during long-running search operations.