:heavy_minus_sign: Remove pull request trigger from Gradle workflow
This commit removes the trigger for pull request on 'main' branch in the GitHub actions Gradle workflow file. This was causing unnecessary runs of the workflow on creating pull requests which resulted in resource wastage. It's decided to run the workflow only on push events to 'main' and 'renovate/**' branches.