:sparkles: Started the implementation of ResultPanel display in Search UI
This commit introduces the feature to display search results in the ResultPanel. The previously empty TODO has been replaced with a complete ResultPanel function in the ResultPanel.kt file. The ResultPanel retrieves the current result state from the controller and forwards it to the newly implemented RealResultPanel Composable. The display format of the search results depends on whether any result is found or not. If no results are found, a "No result found" message is displayed; if results are found, they are displayed in a grid format handled by ResultPanelGrid. A new file ResultGridCell.kt has been created to define each cell of the Result Grid.