:bug: Add padding to ResultPanel and update Text import in ResultGridCell
In the ResultPanel.kt file, modified the RealResultPanel and ResultPanelGrid functions to add padding to layout for a cleaner look. These changes were made to improve the UI and provide better spacing between elements.

In the ResultGridCell.kt file, updated the import statement to use Text from androidx.compose.material3. This was done to be in sync with the latest Material3 design components.