:recycle: Refactor ResultPanel UI component for custom grid cells
Made changes to the `ResultPanel.kt` and `MainScreen.kt` in the song finder-app. This involves refactoring the 'RealResultPanel' component in `ResultPanel.kt` to accept a '@Composable' function as its argument which represents the content of each grid cell. Primarily, this was done to enable greater customizability in terms of how each grid cell in the 'ResultPanel' is displayed.

Corresponding changes were made in `MainScreen.kt` to pass a 'DebugUsedCell' Composable function to the 'RealResultPanel' function as an argument. Also, updated the visibility of 'DebugUsedCell' function in `ResultGridCell.kt` to public to make it accessible in `MainScreen.kt`.