:test_tube: Added TestCellCard.kt for prototyping the layout of a music card
This new file, TestCellCard.kt, implements a concept of music cards for the SongFinder app. It includes UI components like Image, Text, and Icons organized in a Card layout. These music cards will represent individual songs/records in the search result. This update is intended to test and demonstrate the intended design concept for these elements.