:recycle: Refactored hardcoded dp values and replaced them with theme-based dynamic values.
This change reduces code redundancy and improves code readability as well as maintainability. By using theme-based dynamic values for elements like spacing, padding and corner shapes, it's possible to maintain a consistent appearance across different parts of app. With a single place to manage such measurements, we could easily update or change the theme in the future.