:recycle: Refactor ResultPanel to add non-responsive scroll bar
This commit introduces a change in the ResultPanel.kt file to accommodate a scroll bar in the vertical grid of results in the SongFinder app. A Box component and a rememberLazyGridState have been added to maintain the scroll state of the grid. However, this change is currently commented out due to a bug which causes a crash in the application. A potential optimization can be done to ensure that the scroll bar only appears when the window is not being resized but this is yet to be implemented.