:sparkles: Refactored ResultGridCell to use MusicCardTemplate
The ResultGridCell component in the song finder app has been refactored to use the new MusicCardTemplate. This was done to introduce interactive properties to the results grid cell. Now, these grid elements are clickable, triggering a coroutine event. This change was necessary for improving the app's interactivity and user experience.