:heavy_plus_sign: Add our own kamel image config
This commit updates the 'Entrypoint.kt' by integrating a custom Kamel image configuration for handling images in the Song Finder app. This configuration uses the default Kamel settings but with an added 'resourcesFetcher' which is available only on desktop devices. It also defines a custom imageBitmapCacheSize for efficient image handling. This change was necessary to enhance image processing and rendering performance within the app.