:sparkles: Add scroll to top after new search result confirmed
The new ResultGridStateModel has been introduced and added to the SongSearchIntermediateController, ResultPanel and ResultPanelController. When a new search result is confirmed, this model allows the page to automatically scroll back to the top of the result list. This enhancement was made to ensure that users are always provided immediate access to the top results, enhancing user experience by removing the need for manual scrolling when refining search queries.