:sparkles: Completed the lazy thumbnail loading and added caching functionality
This update finalizes the thumbnail rendering feature in the SongFinder app.
Primarily,
we added the ability to lazily load thumbnail images depending on the PV availability.
If a thumbnail fails to load,
the app now automatically tries the next available PV
until a successful thumbnail load occurs
or all options are exhausted.
In the latter case, a failure icon is displayed.
Further,
a caching mechanism has been introduced to speed up thumbnail retrieval.
This cache is evicted every time the user move on to the next song.