:rocket: Get ready for ResultOverridingPanel in MainScreen, the last piece of GUI element before the development is done
Introduced the ResultOverridingPanel to the MainScreen of the Song Finder application. This commit includes the creation of the ResultOverridingPanel file and the ResultOverridingController to handle related business operations, as well as making adjustments to the MainScreen file to incorporate the new Panel. Work is still to be done on the user interface aspect of the ResultOverridingPanel.